Grandes cambios para login y sesiones

+++ src/main/java/com/uah/ismael/portal_formula1/model/repository/RolRepository.java
@@ -3,5 +3,5 @@ package com.uah.ismael.portal_formula1.model.repository;
 import com.uah.ismael.portal_formula1.model.entity.Rol;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolRepository extends JpaRepository<Rol, Long> {
+public interface RolRepository extends JpaRepository<Rol, Integer> {
 }
